Move xml tool parser to separate file
