+++ common/chat.cpp
@@ -153,23 +153,6 @@ struct templates_params {
     bool is_inference = true;
 };
 
-// Sample config:
-// MiniMax-M2 (left): <minimax:tool_call>\n<invoke name="tool-name">\n<parameter name="key">value</parameter>\n...</invoke>\n...</minimax:tool_call>
-// GLM 4.5   (right): <tool_call>function_name\n<arg_key>key</arg_key>\n<arg_value>value</arg_value>\n</tool_call>
-struct xml_tool_call_format {
-    std::string scope_start; // <minimax:tool_call>\n  // \n                      // can be empty
-    std::string tool_start;  // <invoke name=\"        // <tool_call>
-    std::string tool_sep;    // \">\n                  // \n                      // can be empty only for parse_xml_tool_calls
-    std::string key_start;   // <parameter name=\"     // <arg_key>
-    std::string key_val_sep; // \">                    // </arg_key>\n<arg_value>
-    std::string val_end;     // </parameter>\n         // </arg_value>\n
-    std::string tool_end;    // </invoke>\n            // </tool_call>\n
-    std::string scope_end;   // </minimax:tool_call>   //                         // can be empty
-    // Set this if there can be dynamic spaces inside key_val_sep.
-    // e.g. key_val_sep=</arg_key> key_val_sep2=<arg_value> for GLM4.5
-    std::optional<std::string> key_val_sep2 = std::nullopt;
-};
-
 common_chat_tool_choice common_chat_tool_choice_parse_oaicompat(const std::string & tool_choice) {
     if (tool_choice == "auto") {
         return COMMON_CHAT_TOOL_CHOICE_AUTO;
@@ -598,22 +581,6 @@ common_chat_templates_ptr common_chat_templates_init(
             "{%- if false %}");
     }
 
-    // Fix MiniMax-M2 template bug: last_tool_call.name should be tool_call.function.name rather than tool_call.name
-    // TODO: remove this once the template is fixed.
-    if (default_template_src.find("]~!b[") != std::string::npos
-            && default_template_src.find("]~b]") != std::string::npos) {
-        LOG_INF("Detected MiniMax-M2 template , applying automatic fix...\n");
-        if (default_template_src.find("{%- set last_tool_call.name = message.tool_calls[-1].name -%}") != std::string::npos &&
-            default_template_src.find("{%- for tool_call in message.tool_calls -%}") != std::string::npos) {
-            LOG_INF("Detected MiniMax-M2 official template bug: \"last_tool_call.name = message.tool_calls[-1].name\" , applying automatic fix...\n");
-            string_replace_all(default_template_src, "{%- set last_tool_call.name = message.tool_calls[-1].name -%}", "");
-            string_replace_all(default_template_src,
-                "{%- for tool_call in message.tool_calls -%}",
-                "{%- for tool_call in message.tool_calls -%} {%- set last_tool_call.name = tool_call.function.name -%}");
-        }
-        LOG_INF("MiniMax-M2 template fixed\n");
-    }
-
     std::string token_bos = bos_token_override;
     std::string token_eos = eos_token_override;
     bool add_bos = false;
@@ -861,653 +828,6 @@ static std::string apply(
         result = result.substr(0, result.size() - tmpl.eos_token().size());
     }
     return result;
-}
-
-// make a GBNF that accept any strings except those containing any of the forbidden strings.
-inline std::string make_gbnf_excluding(std::vector<std::string> forbids) {
-    constexpr auto charclass_escape = [](unsigned char c) -> std::string {
-        if (c == '\\' || c == ']' || c == '^' || c == '-') {
-            std::string s = "\\";
-            s.push_back((char)c);
-            return s;
-        }
-        if (isprint(c)) {
-            return std::string(1, (char)c);
-        }
-        char buf[16];
-        snprintf(buf, 15, "\\x%02X", c);
-        return std::string(buf);
-    };
-    constexpr auto build_expr = [charclass_escape](auto self, const std::vector<std::string>& forbids, int l, int r, int depth) -> std::string {
-        std::vector<std::pair<unsigned char, std::pair<int,int>>> children;
-        int i = l;
-        while (i < r) {
-            const std::string &s = forbids[i];
-            if ((int)s.size() == depth) {
-                ++i;
-                continue;
-            }
-            unsigned char c = (unsigned char)s[depth];
-            int j = i;
-            while (j < r && (int)forbids[j].size() > depth &&
-                   (unsigned char)forbids[j][depth] == c) {
-                ++j;
-            }
-            children.push_back({c, {i,j}});
-            i = j;
-        }
-        std::vector<std::string> alts;
-        if (!children.empty()) {
-            std::string cls;
-            for (auto &ch : children) cls += charclass_escape(ch.first);
-            alts.push_back(std::string("[^") + cls + "]");
-        }
-        for (auto &ch : children) {
-            std::string childExpr = self(self, forbids, ch.second.first, ch.second.second, depth+1);
-            if (!childExpr.empty()) {
-                std::string quoted_ch = "\"";
-                if (ch.first == '\\') quoted_ch += "\\\\";
-                else if (ch.first == '"') quoted_ch += "\\\"";
-                else if (isprint(ch.first)) quoted_ch.push_back(ch.first);
-                else {
-                    char buf[16];
-                    snprintf(buf, 15, "\\x%02X", ch.first);
-                    quoted_ch += buf;
-                }
-                quoted_ch += "\"";
-                std::string branch = quoted_ch + std::string(" ") + childExpr;
-                alts.push_back(branch);
-            }
-        }
-        if (alts.empty()) return "";
-        std::ostringstream oss;
-        oss << "( ";
-        for (size_t k = 0; k < alts.size(); ++k) {
-            if (k) oss << " | ";
-            oss << alts[k];
-        }
-        oss << " )";
-        return oss.str();
-    };
-    if (forbids.empty()) return "( . )*";
-    sort(forbids.begin(), forbids.end());
-    std::string expr = build_expr(build_expr, forbids, 0, forbids.size(), 0);
-    if (expr.empty()) {
-        std::string cls;
-        for (auto &s : forbids) if (!s.empty()) cls += charclass_escape((unsigned char)s[0]);
-        expr = std::string("( [^") + cls + "] )";
-    }
-    if (forbids.size() == 1)
-        return expr + "*";
-    else
-        return std::string("( ") + expr + " )*";
-}
-
-/**
- * Build grammar for xml-style tool call
- * form.scope_start and form.scope_end can be empty.
- */
-inline void build_grammar_xml_tool_call(common_chat_params & data, const struct templates_params & params, const struct xml_tool_call_format & form) {
-    GGML_ASSERT(!form.tool_start.empty());
-    GGML_ASSERT(!form.tool_sep.empty());
-    GGML_ASSERT(!form.key_start.empty());
-    GGML_ASSERT(!form.val_end.empty());
-    GGML_ASSERT(!form.tool_end.empty());
-
-    std::string key_val_sep = form.key_val_sep;
-    if (form.key_val_sep2) {
-        key_val_sep += "\n";
-        key_val_sep += *form.key_val_sep2;
-    }
-    GGML_ASSERT(!key_val_sep.empty());
-
-    constexpr auto encode_to_safe = [](const std::string &in) {
-        static const char hex[] = "0123456789abcdef";
-        std::string out;
-        out.reserve(in.size() * 4);
-        for (unsigned char uc : in) {
-            if (std::isalnum(uc) || uc == '-') {
-                out.push_back(static_cast<char>(uc));
-            } else {
-                out.push_back('_');
-                out.push_back(hex[(uc >> 4) & 0xF]);
-                out.push_back(hex[uc & 0xF]);
-                out.push_back('_');
-            }
-        }
-        return out;
-    };
-
-    if (params.tools.is_array() && !params.tools.empty()) {
-        data.preserved_tokens.push_back(form.scope_start);
-        data.preserved_tokens.push_back(form.tool_start);
-        data.preserved_tokens.push_back(form.tool_sep);
-        data.preserved_tokens.push_back(form.key_start);
-        data.preserved_tokens.push_back(key_val_sep);
-        data.preserved_tokens.push_back(form.val_end);
-        data.preserved_tokens.push_back(form.tool_end);
-        data.preserved_tokens.push_back(form.scope_end);
-        for (auto &s : data.preserved_tokens) {
-            // s = string_strip(s);
-            s.resize(std::distance(s.begin(), std::find_if(s.rbegin(), s.rend(), [](unsigned char ch) {
-                return !std::isspace(ch);
-            }).base()));
-            size_t start = 0;
-            while (start < s.size() && std::isspace(static_cast<unsigned char>(s[start]))) {
-                ++start;
-            }
-            if (start != 0) {
-                s.erase(0, start);
-            }
-        }
-        data.preserved_tokens.erase(std::remove_if(
-                data.preserved_tokens.begin(),
-                data.preserved_tokens.end(),
-                [](const std::string &s) { return s.size() < 2; }
-        ), data.preserved_tokens.end());
-        std::unordered_set<std::string> seen;
-        seen.reserve(data.preserved_tokens.size());
-        for (auto &s : data.preserved_tokens) {
-            seen.insert(std::move(s));
-        }
-        data.preserved_tokens.assign(
-            std::make_move_iterator(seen.begin()),
-            std::make_move_iterator(seen.end())
-        );
-
-        data.grammar = build_grammar([&](const common_grammar_builder &builder) {
-            std::vector<std::string> tool_rules;
-            foreach_function(params.tools, [&](const json & tool) {
-                const auto & function = tool.at("function");
-                std::string name = function.at("name");
-                std::string name_safe = encode_to_safe(name);
-                auto parameters = function.at("parameters");
-                builder.resolve_refs(parameters);
-
-                std::string param_rules;
-                if (parameters.contains("properties")) {
-                    std::vector<std::string> requiredParameters;
-                    if (parameters.contains("required")) {
-                        try { parameters.at("required").get_to(requiredParameters); }
-                        catch (const std::runtime_error&) {}
-                    }
-                    std::sort(requiredParameters.begin(), requiredParameters.end());
-                    requiredParameters.erase(std::unique(requiredParameters.begin(), requiredParameters.end()), requiredParameters.end());
-                    for (const auto & [key, value] : parameters.at("properties").items()) {
-                        std::string quoted_key = key;
-                        bool required = std::binary_search(requiredParameters.begin(), requiredParameters.end(), key);
-                        if (form.key_start.back() == '"' && key_val_sep[0] == '"') {
-                            quoted_key = gbnf_format_literal(key);
-                            quoted_key = quoted_key.substr(1, quoted_key.size() - 2);
-                        }
-                        if (!required) param_rules += "( ";
-                        param_rules +=
-                                gbnf_format_literal(form.key_start) + " " +
-                                gbnf_format_literal(quoted_key) + " " +
-                                gbnf_format_literal(key_val_sep) + " ";
-                        if (value.contains("type") && value["type"].is_string() && value["type"] == "string") {
-                            param_rules +=
-                                    "( string-arg-val | " +
-                                    builder.add_schema(name_safe + "-arg-" + encode_to_safe(key), value) + " ) ";
-                        } else {
-                            param_rules +=
-                                    builder.add_schema(name_safe + "-arg-" + encode_to_safe(key), value) + " ";
-                        }
-                        param_rules += gbnf_format_literal(form.val_end) + " ";
-                        if (!required) param_rules += ")? ";
-                    }
-                }
-
-                std::string quoted_name = name;
-                if (form.tool_start.back() == '"' && form.tool_sep[0] == '"') {
-                    quoted_name = gbnf_format_literal(name);
-                    quoted_name = quoted_name.substr(1, quoted_name.size() - 2);
-                }
-                tool_rules.push_back(builder.add_rule(name_safe + "-call",
-                        gbnf_format_literal(form.tool_start) + " " +
-                        gbnf_format_literal(quoted_name) + " " +
-                        gbnf_format_literal(form.tool_sep) + " " +
-                        param_rules + " " +
-                        gbnf_format_literal(form.tool_end)
-                ));
-            });
-            builder.add_rule("string-arg-val", make_gbnf_excluding({form.val_end}));
-            builder.add_rule("root", gbnf_format_literal(form.scope_start) + " ( " + string_join(tool_rules, " | ") + " ) " + gbnf_format_literal(form.scope_end));
-        });
-
-        // grammar trigger for tool call
-        data.grammar_lazy = true;
-        data.grammar_triggers.push_back({ COMMON_GRAMMAR_TRIGGER_TYPE_WORD, form.scope_start + form.tool_start });
-    }
-}
-
-/**
- * Parse XML-Style tool call for given xml_tool_call_format. Return false for invalid syntax and get the position untouched.
- * Throws std::runtime_error if there is invalid syntax and cannot recover the original status for common_chat_msg_parser.
- * form.scope_start, form.tool_sep and form.scope_end can be empty.
- */
-inline bool parse_xml_tool_calls(common_chat_msg_parser & builder, const struct xml_tool_call_format & form) {
-    GGML_ASSERT(!form.tool_start.empty());
-    GGML_ASSERT(!form.key_start.empty());
-    GGML_ASSERT(!form.key_val_sep.empty());
-    GGML_ASSERT(!form.val_end.empty());
-    GGML_ASSERT(!form.tool_end.empty());
-
-    constexpr auto all_space = [] (auto &str) {
-        return std::all_of(str.begin(), str.end(), [](unsigned char ch) { return std::isspace(ch); });
-    };
-    // Helper to choose return false or throw error
-    constexpr auto return_error = [](common_chat_msg_parser & builder, auto &start_pos, const bool &recovery) {
-        LOG_DBG("Failed to parse XML-Style tool call at position: %s\n", gbnf_format_literal(builder.consume_rest().substr(0, 20)).c_str());
-        if (recovery) {
-            builder.move_to(start_pos);
-            return false;
-        } else throw std::runtime_error("Tool call parsing failed with unrecoverable errors. Try using a grammar to constrain the model’s output.");
-    };
-    // Drop substring from needle to end from a JSON
-    constexpr auto partial_json = [](std::string &json_str, std::string_view needle = "XML_TOOL_CALL_PARTIAL_FLAG") {
-        auto pos = json_str.rfind(needle);
-        if (pos == std::string::npos) {
-            return false;
-        }
-        for (auto i = pos + needle.size(); i < json_str.size(); ++i) {
-            unsigned char ch = static_cast<unsigned char>(json_str[i]);
-            if (ch != '\'' && ch != '"' && ch != '}' && ch != ':' && !std::isspace(ch)) {
-                return false;
-            }
-        }
-        if (pos != 0 && json_str[pos - 1] == '"') {
-            --pos;
-        }
-        json_str.resize(pos);
-        return true;
-    };
-    // Helper to generate a partial argument JSON
-    constexpr auto gen_partial_json = [partial_json](auto &&set_partial_arg, auto &&arguments, auto &&builder, auto &&function_name) {
-        std::forward<decltype(set_partial_arg)>(set_partial_arg)(std::forward<decltype(builder)>(builder).consume_rest(), "XML_TOOL_CALL_PARTIAL_FLAG");
-        auto tool_str = std::forward<decltype(arguments)>(arguments).dump();
-        if (partial_json(tool_str)) {
-            if (std::forward<decltype(builder)>(builder).add_tool_call(std::forward<decltype(function_name)>(function_name), "", tool_str)) {
-                return;
-            }
-        }
-        LOG_DBG("Failed to parse partial XML-Style tool call, fallback to non-partial: %s\n", tool_str.c_str());
-    };
-
-    bool recovery = true;
-    const auto start_pos = builder.pos();
-    if (!all_space(form.scope_start) && !builder.try_consume_literal(form.scope_start)) return false;
-    while (auto tc = builder.try_find_literal(form.tool_start)) {
-        if (!all_space(tc->prelude)) {
-            LOG_DBG("Failed to parse XML-Style tool call: Expected %s, but found %s\n",
-                    gbnf_format_literal(form.tool_start).c_str(),
-                    gbnf_format_literal(tc->prelude).c_str()
-            );
-            return return_error(builder, start_pos, recovery);
-        }
-
-        // Find tool name
-        auto func_name = builder.try_find_literal(all_space(form.tool_sep) ? form.key_start : form.tool_sep);
-        if (!func_name) {
-            func_name = builder.try_find_literal(form.tool_end);
-        }
-        if (!func_name) {
-            // Partial tool name not supported
-            throw common_chat_msg_partial_exception("incomplete tool_call");
-        }
-        // If the model generate multiple tool call and the first tool call has no argument
-        if (func_name->prelude.find(form.tool_end) != std::string::npos) {
-            builder.move_back(func_name->prelude.size() + form.tool_end.size());
-            func_name = builder.try_find_literal(form.tool_end);
-        }
-
-        // Parse tool name
-        builder.move_to(all_space(form.tool_sep) ? func_name->groups[0].begin : func_name->groups[0].end);
-        std::string function_name = string_strip(func_name->prelude);
-
-        // Argument JSON
-        json arguments = json::object();
-
-        // Helper to generate a partial argument JSON
-        const auto gen_partial_args = [&](auto &&set_partial_arg) {
-            gen_partial_json(std::forward<decltype(set_partial_arg)>(set_partial_arg), arguments, builder, function_name);
-        };
-
-        // Parse all arg_key/arg_value pairs
-        while (auto tc = builder.try_find_literal(form.key_start)) {
-            if (tc->groups[0].end - tc->groups[0].begin != form.key_start.size()) {
-                auto tool_call_arg = arguments.dump();
-                if (tool_call_arg.size() != 0 && tool_call_arg[tool_call_arg.size() - 1] == '}') {
-                    tool_call_arg.resize(tool_call_arg.size() - 1);
-                }
-                builder.add_tool_call(function_name, "", tool_call_arg);
-                throw common_chat_msg_partial_exception("Partial literal: " + gbnf_format_literal(form.key_start));
-            }
-            if (!all_space(tc->prelude)) {
-                LOG_DBG("Failed to parse XML-Style tool call: Expected %s, but found %s\n",
-                        gbnf_format_literal(form.key_start).c_str(),
-                        gbnf_format_literal(tc->prelude).c_str()
-                );
-                return return_error(builder, start_pos, recovery);
-            }
-
-            // Parse arg_key
-            auto key_res = builder.try_find_literal(form.key_val_sep);
-            if (!key_res) {
-                gen_partial_args([&](auto &&rest, auto &&needle) {arguments[rest + needle] = "";});
-                throw common_chat_msg_partial_exception("Expected " + gbnf_format_literal(form.key_val_sep) + " after " + gbnf_format_literal(form.key_start));
-            }
-            if (key_res->groups[0].end - key_res->groups[0].begin != form.key_val_sep.size()) {
-                gen_partial_args([&](auto &&, auto &&needle) {arguments[key_res->prelude + needle] = "";});
-                throw common_chat_msg_partial_exception("Partial literal: " + gbnf_format_literal(form.key_val_sep));
-            }
-            auto &key = key_res->prelude;
-            recovery = false;
-
-            // Parse arg_value
-            if (form.key_val_sep2) {
-                if (auto tc = builder.try_find_literal(*form.key_val_sep2)) {
-                    if (tc->groups[0].end - tc->groups[0].begin != form.key_val_sep2->size()) {
-                        gen_partial_args([&](auto &&, auto &&needle) {arguments[key] = needle;});
-                        throw common_chat_msg_partial_exception("Partial literal: " + gbnf_format_literal(*form.key_val_sep2));
-                    }
-                    if (!all_space(tc->prelude)) {
-                        LOG_DBG("Failed to parse XML-Style tool call: Unexcepted %s between %s and %s\n",
-                                gbnf_format_literal(tc->prelude).c_str(),
-                                gbnf_format_literal(form.key_val_sep).c_str(),
-                                gbnf_format_literal(*form.key_val_sep2).c_str()
-                        );
-                        return return_error(builder, start_pos, false);
-                    }
-                } else {
-                    gen_partial_args([&](auto &&, auto &&needle) {arguments[key] = needle;});
-                    throw common_chat_msg_partial_exception("Expected " + gbnf_format_literal(*form.key_val_sep2) + " after " + gbnf_format_literal(form.key_val_sep));
-                }
-            }
-            auto val_start = builder.pos();
-
-            // Test if arg_val is a partial JSON
-            std::optional<common_json> value_json = std::nullopt;
-            try { value_json = builder.try_consume_json(); }
-            catch (const std::runtime_error&) { builder.move_to(val_start); }
-
-            // If it is a JSON and followed by </arg_value>, parse as json
-            // cannot support streaming because it may be a plain text starting with JSON
-            if (value_json) {
-                auto tmp_pos = builder.pos();
-                builder.consume_spaces();
-                if (builder.pos() == builder.input().size()) {
-                    gen_partial_args([&](auto &&, auto &&needle) {arguments[key] = needle;});
-                    LOG_DBG("Possible JSON arg_value: %s\n", value_json->json.dump().c_str());
-                    throw common_chat_msg_partial_exception("JSON arg_value detected. Waiting for more tokens for validations.");
-                }
-                builder.move_to(tmp_pos);
-                auto tc = builder.try_find_literal(form.val_end);
-                if (tc && value_json->healing_marker.marker.empty()) {
-                    if (tc->groups[0].end - tc->groups[0].begin != form.val_end.size()) {
-                        gen_partial_args([&](auto &&, auto &&needle) {arguments[key] = needle;});
-                        LOG_DBG("Possible terminated JSON arg_value: %s\n", value_json->json.dump().c_str());
-                        throw common_chat_msg_partial_exception("Partial literal: " + gbnf_format_literal(form.val_end));
-                    }
-                    if (all_space(tc->prelude)) {
-                        arguments[key] = value_json->json;
-                    }
-                } else builder.move_to(val_start);
-            }
-
-            // If not, parse as plain text
-            if (val_start == builder.pos()) {
-                if (auto value_plain = builder.try_find_literal(form.val_end)) {
-                    if (value_plain->groups[0].end - value_plain->groups[0].begin != form.val_end.size()) {
-                        gen_partial_args([&](auto &&, auto &&needle) {arguments[key] = value_plain->prelude + needle;});
-                        throw common_chat_msg_partial_exception(
-                                "Expected " + gbnf_format_literal(form.val_end) +
-                                " after " + gbnf_format_literal(form.key_val_sep) +
-                                (form.key_val_sep2 ? " " + gbnf_format_literal(*form.key_val_sep2) : "")
-                        );
-                    }
-                    arguments[key] = value_plain->prelude;
-                } else {
-                    gen_partial_args([&](auto &&rest, auto &&needle) {arguments[key] = rest + needle;});
-                    throw common_chat_msg_partial_exception(
-                            "Expected " + gbnf_format_literal(form.val_end) +
-                            " after " + gbnf_format_literal(form.key_val_sep) +
-                            (form.key_val_sep2 ? " " + gbnf_format_literal(*form.key_val_sep2) : "")
-                    );
-                }
-            }
-        }
-
-        // Consume closing tag
-        if (auto tc = builder.try_find_literal(form.tool_end)) {
-            if (!all_space(tc->prelude)) {
-                LOG_DBG("Failed to parse XML-Style tool call: Expected %s, but found %s\n",
-                        gbnf_format_literal(form.tool_end).c_str(),
-                        gbnf_format_literal(tc->prelude).c_str()
-                );
-                return return_error(builder, start_pos, recovery);
-            }
-            if (tc->groups[0].end - tc->groups[0].begin == form.tool_end.size()) {
-                // Add the parsed tool call
-                if (!builder.add_tool_call(function_name, "", arguments.dump())) {
-                    throw common_chat_msg_partial_exception("Failed to add XML-Style tool call");
-                }
-                recovery = false;
-                continue;
-            }
-        }
-
-        auto tool_call_arg = arguments.dump();
-        if (tool_call_arg.size() != 0 && tool_call_arg[tool_call_arg.size() - 1] == '}') {
-            tool_call_arg.resize(tool_call_arg.size() - 1);
-        }
-        builder.add_tool_call(function_name, "", tool_call_arg);
-        throw common_chat_msg_partial_exception("Expected " + gbnf_format_literal(form.tool_end) + " after " + gbnf_format_literal(form.val_end));
-    }
-    if (auto tc = builder.try_find_literal(form.scope_end)) {
-        if (!all_space(tc->prelude)) {
-            LOG_DBG("Failed to parse XML-Style tool call: Expected %s, but found %s\n",
-                    gbnf_format_literal(form.scope_end).c_str(),
-                    gbnf_format_literal(tc->prelude).c_str()
-            );
-            return return_error(builder, start_pos, recovery);
-        }
-    } else {
-        if (all_space(form.scope_end)) return true;
-        builder.consume_spaces();
-        if (builder.pos() == builder.input().size())
-            throw common_chat_msg_partial_exception("incomplete tool calls");
-        LOG_DBG("Failed to parse XML-Style tool call: Expected %s, but found %s\n",
-                gbnf_format_literal(form.scope_end).c_str(),
-                gbnf_format_literal(builder.consume_rest()).c_str()
-        );
-        return return_error(builder, start_pos, recovery);
-    }
-
-    return true;
-}
-
-// Parse content uses reasoning and XML-Style tool call
-inline void parse_msg_with_xml_tool_calls(common_chat_msg_parser & builder, const struct xml_tool_call_format & form, const std::string & start_think = "<think>", const std::string & end_think = "</think>") {
-    constexpr auto rstrip = [](std::string &s) {
-        s.resize(std::distance(s.begin(), std::find_if(s.rbegin(), s.rend(), [](unsigned char ch) { return !std::isspace(ch); }).base()));
-    };
-    // Erase substring from l to r, along with additional spaces nearby
-    constexpr auto erase_spaces = [](auto &str, size_t l, size_t r) {
-        while (/* l > -1 && */ --l < str.size() && std::isspace(static_cast<unsigned char>(str[l])));
-        ++l;
-        while (++r < str.size() && std::isspace(static_cast<unsigned char>(str[r])));
-        if (l < r) str[l] = '\n';
-        if (l + 1 < r) str[l + 1] = '\n';
-        if (l != 0) l += 2;
-        str.erase(l, r - l);
-        return l;
-    };
-    // Handle unclosed </think> from content
-    constexpr auto filter_unclosed_think = [erase_spaces](auto &content, auto &&builder, const std::string &end_think) {
-        auto &syntax = std::forward<decltype(builder)>(builder).syntax();
-        if (syntax.reasoning_format == COMMON_REASONING_FORMAT_NONE || syntax.reasoning_in_content) return;
-        if (auto pos = content.rfind(end_think); pos != std::string::npos) {
-            // delete all </think> token
-            while (pos != std::string::npos) {
-                pos = erase_spaces(content, pos, pos + end_think.size() - 1);
-                pos = content.rfind(end_think, pos);
-            }
-        }
-    };
-    // Escape string literal to regex that match the literal
-    constexpr auto escape_regex = [](const std::string &s) {
-        // Characters that are regex metacharacters in ECMAScript grammar:
-        const std::string meta = R"(\^$.*+?()[]{}|)"; // backslash included
-        std::string out;
-        out.reserve(s.size() * 3 + 2); // rough reserve
-        for (unsigned char uc : s) {
-            // Printable ASCII range we allow to remain unescaped: letters, digits, underscore
-            if ((uc >= '0' && uc <= '9') ||
-                (uc >= 'A' && uc <= 'Z') ||
-                (uc >= 'a' && uc <= 'z') ||
-                uc == '_') {
-                out.push_back(static_cast<char>(uc));
-            } else if (meta.find(static_cast<char>(uc)) != std::string::npos) {
-                // regex metacharacter -> escape with backslash
-                out.push_back('\\');
-                out.push_back(static_cast<char>(uc));
-            } else if (uc >= 0x20 && uc <= 0x7E) {
-                // other printable ASCII (space, punctuation not in meta) -> keep
-                out.push_back(static_cast<char>(uc));
-            } else {
-                switch (uc) {
-                    case '\0': out += "\\0"; break; // NUL
-                    case '\a': out += "\\a"; break; // Bell (0x07)
-                    case '\b': out += "\\b"; break; // Backspace (0x08)
-                    case '\f': out += "\\f"; break; // Formfeed (0x0C)
-                    case '\n': out += "\\n"; break; // Linefeed (0x0A)
-                    case '\r': out += "\\r"; break; // Carriage return (0x0D)
-                    case '\t': out += "\\t"; break; // Horizontal tab (0x09)
-                    case '\v': out += "\\v"; break; // Vertical tab (0x0B)
-                    default: {
-                        // It seems the current partial-regex implementation doesn’t support this form and will silently fail
-                        // TODO: delete this when \xHH is supported by partial-regex
-                        throw std::runtime_error("Cannot escape non-printable or non-ASCII byte for string: " + gbnf_format_literal(s));
-                        // Non-printable or non-ASCII byte: use \xHH
-                        std::ostringstream oss;
-                        oss << "\\x" << std::hex << std::uppercase << std::setw(2) << std::setfill('0') << int(uc);
-                        out += oss.str();
-                    }
-                }
-            }
-        }
-        return out;
-    };
-
-    const common_regex tool_call_start_regex(escape_regex(form.scope_start) + "\\s*" + escape_regex(form.tool_start));
-    LOG_DBG("Regex for tool start: %s\n", (escape_regex(form.scope_start) + "\\s*" + escape_regex(form.tool_start)).c_str());
-
-    // Parse content
-    bool reasoning_unclosed = builder.syntax().thinking_forced_open;
-    std::string unclosed_reasoning_content("");
-    for (;;) {
-        auto tc = builder.try_find_regex(tool_call_start_regex, std::string::npos, false);
-        std::string content;
-        std::string tool_call_start;
-
-        if (tc) {
-            content = std::move(tc->prelude);
-            tool_call_start = builder.str(tc->groups[0]);
-            LOG_DBG("Matched tool start: %s\n", gbnf_format_literal(tool_call_start).c_str());
-        } else {
-            content = builder.consume_rest();
-        }
-
-        // Handle unclosed think block
-        if (reasoning_unclosed) {
-            if (auto pos = content.find(end_think); pos == std::string::npos && builder.pos() != builder.input().size()) {
-                unclosed_reasoning_content += content + tool_call_start;
-                continue;
-            } else {
-                std::string reasoning_content;
-                if (pos == std::string::npos) {
-                    reasoning_content = std::move(content);
-                } else {
-                    reasoning_content = content.substr(0, pos);
-                    content.erase(0, pos + end_think.size());
-                }
-                if (builder.syntax().reasoning_format == COMMON_REASONING_FORMAT_NONE || builder.syntax().reasoning_in_content) {
-                    if (builder.result().content.size() != 0) {
-                        builder.add_content("\n\n");
-                    }
-                    builder.add_content(start_think);
-                    builder.add_content(unclosed_reasoning_content);
-                    builder.add_content(reasoning_content);
-                    if (builder.pos() != builder.input().size() || std::any_of(content.begin(), content.end(), [](unsigned char c) { return !std::isspace(c); }))
-                        builder.add_content(end_think);
-                } else {
-                    builder.add_reasoning_content(unclosed_reasoning_content);
-                    builder.add_reasoning_content(reasoning_content);
-                }
-                unclosed_reasoning_content.clear();
-                reasoning_unclosed = false;
-            }
-        }
-
-        // Handle multiple think block
-        bool toolcall_in_think = false;
-        for (auto think_start = content.rfind(start_think); think_start != std::string::npos; think_start = content.rfind(start_think, think_start - 1)) {
-            if (auto think_end = content.find(end_think, think_start + start_think.size()); think_end != std::string::npos) {
-                if (builder.syntax().reasoning_format != COMMON_REASONING_FORMAT_NONE && !builder.syntax().reasoning_in_content) {
-                    auto reasoning_content = content.substr(think_start + start_think.size(), think_end - think_start - start_think.size());
-                    builder.add_reasoning_content(reasoning_content);
-                    think_start = erase_spaces(content, think_start, think_end + end_think.size() - 1);
-                }
-            } else {
-                // This <tool_call> start is in thinking block, skip this tool call
-                auto pos = think_start + start_think.size();
-                unclosed_reasoning_content = content.substr(pos) + tool_call_start;
-                reasoning_unclosed = true;
-                content.resize(think_start);
-                toolcall_in_think = true;
-            }
-        }
-        rstrip(content);
-
-        // Handle unclosed </think> token
-        filter_unclosed_think(content, builder, end_think);
-
-        // Strip if needed
-        if (content.size() > 0 && std::isspace(static_cast<unsigned char>(content[0]))) {
-            content = string_strip(content);
-        }
-
-        // Add content
-        if (content.size() != 0) {
-            // If there are multiple content blocks
-            if (builder.result().content.size() != 0) {
-                builder.add_content("\n\n");
-            }
-            builder.add_content(content);
-        }
-
-        // This <tool_call> start is in thinking block, skip this tool call
-        if (toolcall_in_think) {
-            continue;
-        }
-
-        // There is no tool call and all content is parsed
-        if (!tc) {
-            GGML_ASSERT(builder.pos() == builder.input().size());
-            GGML_ASSERT(unclosed_reasoning_content.empty());
-            GGML_ASSERT(!reasoning_unclosed);
-            break;
-        }
-
-        builder.move_to(tc->groups[0].begin);
-        if (!parse_xml_tool_calls(builder, form)) {
-            static const common_regex next_char_regex(".");
-            auto c = builder.str(builder.consume_regex(next_char_regex).groups[0]);
-            rstrip(c);
-            builder.add_content(c);
-        }
-    }
 }
 
 static common_chat_params common_chat_params_init_generic(const common_chat_template & tmpl, const struct templates_params & inputs) {
@@ -2538,7 +1858,7 @@ static common_chat_params common_chat_params_init_minimax_m2(const common_chat_t
         /* form.tool_end    = */ "</invoke>\n",
         /* form.scope_end   = */ "</minimax:tool_call>",
     };
-    build_grammar_xml_tool_call(data, params, form);
+    build_grammar_xml_tool_call(data, params.tools, form);
 
     return data;
 }
@@ -2554,7 +1874,7 @@ static void common_chat_parse_minimax_m2(common_chat_msg_parser & builder) {
         /* form.tool_end    = */ "</invoke>",
         /* form.scope_end   = */ "</minimax:tool_call>",
     };
-    parse_msg_with_xml_tool_calls(builder, form, "<think>", "</think>");
+    builder.consume_reasoning_with_xml_tool_calls(form, "<think>", "</think>");
 }
 
 static common_chat_params common_chat_params_init_gpt_oss(const common_chat_template & tmpl, const struct templates_params & inputs) {
@@ -2875,7 +2195,7 @@ static common_chat_params common_chat_params_init_glm_4_5(const common_chat_temp
         /* form.tool_end    = */ "</tool_call>\n",
         /* form.scope_end   = */ "",
     };
-    build_grammar_xml_tool_call(data, inputs, form);
+    build_grammar_xml_tool_call(data, inputs.tools, form);
 
     data.prompt = prompt;
     data.format = COMMON_CHAT_FORMAT_GLM_4_5;
@@ -2894,7 +2214,7 @@ static void common_chat_parse_glm_4_5(common_chat_msg_parser & builder) {
         /* form.scope_end    = */ "",
         /* form.key_val_sep2 = */ "<arg_value>",
     };
-    parse_msg_with_xml_tool_calls(builder, form, "<think>", "</think>");
+    builder.consume_reasoning_with_xml_tool_calls(form, "<think>", "</think>");
 }
 
 static common_chat_params common_chat_params_init_firefunction_v2(const common_chat_template & tmpl, const struct templates_params & inputs) {
@@ -3580,7 +2900,7 @@ static void common_chat_parse_seed_oss(common_chat_msg_parser & builder) {
         /* form.tool_end    = */ "</function>",
         /* form.scope_end   = */ "</seed:tool_call>",
     };
-    parse_msg_with_xml_tool_calls(builder, form, "<seed:think>", "</seed:think>");
+    builder.consume_reasoning_with_xml_tool_calls(form, "<seed:think>", "</seed:think>");
 }
 
 static common_chat_params common_chat_params_init_without_tools(const common_chat_template & tmpl, const struct templates_params & inputs) {
